refactor(#87): PixelStatRow 공용화
